feat(library): universele "+"-knop in de header voor wet, werkdocument en leden
Vervangt de losse "Wet toevoegen"-plus onder de wettenlijst door één "+"-knop
in de header, links van de trajecten/account-bar. Het menu opent:

- Wet toevoegen: nu een sheet met een segmented control Zoeken/Uploaden, in
  plaats van de gekopieerde search-popover-combobox (AddLawSheet, vervangt
  AddLawPopover). De upload-route zit in de Uploaden-tab.
- Werkdocument toevoegen: submenu met Nieuw document en PDF/DOCX uploaden.
- Leden uitnodigen: de uit TrajectMembersPane gelichte InviteMembersSheet,
  overal te openen zonder eerst naar de leden-pagina te navigeren.

Uitgelogd blijft de knop staan als ontdekpunt en opent een popover die
uitnodigt in te loggen (zelfde patroon als de trajecten-knop).

De header stuurt intenties via een nieuwe useAddActions-composable; LibraryView
en de leden-pane vangen ze op (LibraryView is altijd gemount binnen een traject;
de leden-pane herlaadt op een membersChanged-signaal na een invite). Traject-
gescoped, met "…" voor sheet-openers en een chevron voor het submenu.
